requirements.txt
##### use latest version
pillow
prettytable
geopy